msg sending implemented
msg_issue_xxx functions declared, implemented and used
See that pp_get_tstamp now returns a TimeInternal struct.
Actually timestamp struct might be defined outside of ieee_1588_types.h
file, since it is not defined in the spec. For the moment, I leave
it there, but in the future it must be redefined for our purposes (and
in standard_c_notation)