no_incoming_msg label introduce to skip switch when plen=0
This patch is due to the fact that the code did not correctly handle
the situation of "no incoming message". Since messageType 0 is a SYNC
message type, in those cases an empty message was tried to be handled as a SYNC
message, and led to FAULTY state. This patch seemed to me the least
invasive with respect to the current code.