proto-ext-whiterabbit development starting point
This commit simply copies everything from proto-standard to
proto-ext-whiterabbit. Actually, this is not an extension yet, since it is
quite hard to separate what is "modified" from what is "extended". So we
will have a separate directory for white rabbit, develop it as a separate
and independent plugin, and, a the end, split what is an extension of
proto-standard.

Notice that, with the current commit, whiterabbit is set as default in the
Makefile, so that you do not need to make PROTO_EXT=whiterabbit, but simply
make. In order to compile proto-standard, just set empty PROTO_EXT:
make PROTO_EXT=