wr state machine functions skeleton
This patch introduces the state machine for white rabbit.
It is intended as a flat state machine, so that the application state is
univocally determined by its ppsi->state variable. Internally, each state
function saves the Port Dataset, keeping the double level of states: portState
for general state, and wrPortState as White Rabbit FSM. This permits to be
logically coherent with the specification document (which defines an internal
WR state machine) and to have a simplified approach in the source code.