arch-spec/spec-socket.c: initialize both EVT and GEN socket with our spec sock
PTP protocol defines two kind of sockets, "event" and "general". In spec, and
in raw socket as a general rule, this distinction is not present, but ppsi
makes use of both of them. So, with this patch, we use the same socket for
both. This solves a bug in the initialization of clockIdentity with the MAC
addr.
See proto-ext-whiterabbit/state-initializing.c, from line 36:

	id = DSDEF(ppi)->clockIdentity;
	mac = NP(ppi)->ch[PP_NP_GEN].addr;
	id[0] = mac[0];
	id[1] = mac[1];
	id[2] = mac[2];
	id[3] = 0xff;

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>