diag: bugfix for arch-gnu-linux
pp_printf calls puts, but that's the "embedded" puts, the one that
doens't add a trailing newline. So to avoid extra lines and split
messages in the diagnostic output we need to link the fake puts, that
calls pp_puts instead (in turn, pp_puts for arch-gnu-linux relies on
fputs, which doesn't add the trailing newline).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>