diag: documented status and prepared changes
This commit makes no technical changesin the output, but it creates
<ppsi/diag-macros.h> (always includes) to split out all the
diag-related stuff.  The macros in there are the same as in previous
commits, but they are documented.

This also introduces pp_verbose_messages (which is 1 or 0) to allow
the various "#ifdef VERB_LOG_MSGS" to become language if() instead.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>