WR: remove used fields in TimeInternal
The three "raw" fields claim to exist for debugging, but since we
have several TimeInternal structures around they cost quite some space.
The fields were only copied to TimeInternal from network frames,
and there is not explicit code to dump them.

Thus, this commit removes them but leaves the code in place, so it can
be re-enabled if the need arises. (I use the ugly "#if 0").

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>