massive: remove pdelay material.
pdelay (peer delay) is a configuration option for a PTP domain. IT is
a huge mass of code, which we never actually used and is most likely
buggy.

This commits remove all related code, and configures our PTP engine as
end-to-end.  I doubt this commit can be reverted in the future, but at
least it leaves traces about what was there and is not there any more.

Something about pdelay is still there: the message names, one field in
a structure and one configuration variable, to ease recovery in the
future.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>