kill MSG_SEND_AND_RET, use function instead
This removes the unmanageable MSG_SEND_AND_RET macro, turning it to a
normal function (inline, at the time being, to avoid moving from the
header to a different place).

The function is called __send_and_log because it logs too, but I'd better
move logging to the actual send and recv methods.

BTW: the message "-> FAULTY" (now removed) was wrong, because not all
errors of this function result in FAULTY state in the fsm.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>