kill timer: implement timeout
The timer implementatin was incredibly complex.  Now there is no such
thing as a timer in ptp. We only have timeouts.

A timeout is a point in time. We calculate; we can be earlier; we
can be later. The only thing which is arch-specific is calculating.

This replaces every "timer" operation with the equivalent "timeout"
one. All stuff is renamed to avoid confusion between "time" and
"timer".

Unfortunately, the resulting program as of this commit is not working.
the pdelay_request timeout happens continuously, I'm sure I've exposed
a latent bug, which I'm now going to fix.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>