timeout: add pp_timeout_z and centralized logging
By zeroing the timeout whenever it expires, a number of
state files get simplified a little. So pp_timeout_z() returns
1 or 0 like pp_timeout(), but also clears the timeout when it is
expired.

Moreover, when a timeout expires a message is printed if the
build-time option pp_verbose_time is set (and run-time verbosity is at
level 2). In this way all timeouts are reported, or none of them.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>