timeout: randomize intervals as required
The frames should be sent with some dispersion around the timeout
point.  The new function calculates the correct dispersion, but we
need to change the setting of next_delay in order to actually spread
the distribution of announce and other messages.

Also, this commit doesn't spread all message times, some work on the
field is still needed.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>