general: fix sent_seq_id in ppi structure
The structure was separately allocated with a size of 16.  This avoids
the alloc and makes the array the right size from the enum of message
types. We thus a remove a number of magic "16" in the code.
wrpc-sw must be updated too, in order to use this commit.

There's still a problem: the code claims the number to be "last sent",
but actually the value is incremented after sending. I'll need to pass
again on this thing.

ps: I changed name from "sent_seq_id" to "sent_seq" to the compiler
complains for previous users, otherwise undetectable.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>