arch-bare-*: unify common code
This unifies all bare-linux common code. It's an out-of-context commit
(as I'm not adding bare-arm today) but I prefer to avoid doing the same
change twice in things that are going to change these days.

BTW: this commit introduces no changes in the binary files generated,
with the exception of a small difference between i386 and x86-64 in
the receive buffer.  This keeps the x86-64 version but i386 was not
working anyways (and still doesn't work).  That bug fix is in my queue,
but I've other things to do first with ppsi.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>