general: time operations are now per-instance
This commit moves the time operations inside the instance.  It
allows to have different time operations for each instance, which will be
useful in the future. For example, one ethernet card may White Rabbit,
and another may not be.

This commit, like moving network operations inside ppi, requires a
change in wrpc-sw.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>