general: net: rename some fields, fix alignment
This changes some things to make code more readable (I really didn't
know what the tmp_header was: now it is the "received_ptp_header".
The commit simplifies allocation, because the tx and rx buffers are
now part of ppi itself.

We now have tx_ptp and tx_frame, both pointing within tx_buffer (and
the same for rx). The pointers are the protocol side (aligned) and the
argument passed to send/recv (which may be not be 4-aligned).

At this point nothing changes, i.e. both gnu-linux flavours work and
bare-i386 is not able to received.