fix frames offsets (now bare works)
This commits removes all the additions and subtractions of
the protocol offset. Now all recv and send functions, receive
the pointer to the actual data being sent, be it the ptp payload
or the transport header.

This simplification (though based on the two internal pointers within
ppsi -- se previous commit) fixed a long standing bug with bare
architectures, that didn't receive the frames due to some related
problem in offset calculation.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>