fsm.c: bugfix on frame error
fsm.c corrently decodes the received frame, and error management is
built in (even if code in msg.c still needs to be audited to be more
careful about rogue frames).  However, when a received frame is in
error, fsm.c cannot return to the caller because it doesn't know the
current delay. Thus, zero the frame and call the current status, so
we have back a recalculated ppi->next_delay.

Meanwhile, mask with pp_verbose_frames the diagnostic message for
received frames -- and be more careful with accessing it's data.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>