general: add ppi argument to all time operations
Unfortunately, we need to pass ppi to the time operations, so we
can have per-instance logging options. This is the second time I found
the need for such extra argument: the first time I worked it around, now
I surrender.

Actually, I don't find it unlikely for a multi-instance process to
spit a lot of diagnostics about timestamps. A developer may need to
filter messages according to who is tiemstamping/timing out etc.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>