general (for bmc.c): turn ClockIdentity into a structure
This "type" was the only type represented as an array instead of
being a structure. This leads to inconsistent use: if you memcpy
the other ones you need "&", but not here.

So this commit uses an encompassing structure, to have ClockIdentity
behave like other composite fields.  The commit has no effect at all
on the generated binary files for wrpc but it unexpectedly reduces the
size of the output on the PC. Anyways, the commit is trivial, and all
changes can be easily verified to have no effect.

While this increases entropy, the reasoning is that I want to assign
structures instead of running zillions of memcpy functions, with the
hairy PP_CLOCK_IDENTITY_LENGTH.  So later commits, based on this one,
reduce entropy.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>