announce data structure: rename wrFlags to ext_specific
Since WR only needs one byte of information in the announce structure,
we can make is generic by calling it differently.

This changes only the name (and the size: it is now unsigned long),
but has no effect on the marshalling/unmarshalling of such data.

I'm touching this because I'm working on announce messages (as related
to foreign masters).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>