bmc: some simplifications more
A large number of functions here got hdr and ann, but struct
pp_frgn_master includes such fields, and most callers passed hdr and
ann from the same structure.

So I chose to pass a foreign master instead of the two components, but
I'm still unable to remove the two full structures inside until I
clean up the msg.c mess.   So, at the end, this is not really a step
forward, but it helps such later step.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>