kill ppi->msg_tmp
There is no need for a "msg_tmp" in the pp-instance. All uses are
really temporary to a function, so we can use local variables instead.

(This is not the same as what was called tmp_header, which was
referenced all over message processing, so it still exists, though
renamed).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>