kill ppi->record_update
The code, when receiving announce, was setting a flag in order to run
bmc at the next iteration, when that happened, the frame just received
was discarded in favour for the programmed state update.

This commits remove some fields and one helper function, so the new
announce is immediately processed and no frame is lost.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>