bmc: simplify dataset_cmp again
This simplifies the dataset_cmp a little more, and adds some comment.
Nothing changes in the outcome of this commit but I'm unsure about the
code (i.e about the original code).

The specification says to compare *port* identities, but we are
comparing clock identities instead. This is the same in the original
ptpd, so if it is wrong it has always been.

We'll be able to see what happens when checking with several masters,
after porting ppsi to multi-link setup.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>