tools/dump-funcs: make dumpstruct less generic
This commit removes the "dest" argument to dumpstruct(), which now
uses printf() and not fprintf(dest,).  I plan to use these functions
withing PPSi itself if the "frames" diagnostic level is 2, and I'd
better avoid to hack around fprintf.

ptpdump is not changed, as it already used dumpstruct() with stdout
as destination.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>