bugfix: ieee1588_types.h: remove unused data
In commit 7c14152c I removed some typedef, in my quest for getting rid
of them.  But I forgot to remove the type name, so we got "TLV" and
similar global variables in the object files.  This fixes the buglet,
then I add -fno-common to prevent this for happening againg.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>