general: rename fields within ppsi (no technical change)
This renames sync_receive_time to t2, delay_req_send_time to t3
and delay_req_receive_time to t4 (there is "t1" field in ppi yet).

Such naming is what the official documents use everywhere, so it's
better to respect that naming, that everybody in the field is familiar
with. Besides, they are shorter and they can be used unchanged for
pdelay operations (the previous pdelay implementation we inherited
from ptpd, used a separate set of fields that are not really needed).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>