bugfix: correctionField conversion
I initially looked at int64_to_TimeInternal to swap arguments, since
to_TimeInternal as timeinternal (the output) as first argument, and
they appeared in pairs.  However, int64_to_TimeInternal is only used
for correction fields, and they are scaled nanoseconds (page 126), not
nanoseconds.

Thus, this commit fixes the conversion and renames the function to
"cField_to_TimeInternal" to better tell the reader what it is.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>