bugfix: correctionField use
For two-step masters, the code was adding the correction fields of
sync and follow-up.  This is wrong, since the follow-up message only
carries the sender's timestamp for sync.  So, the only correction to
consider is the one in sync (which accounts for transparent clocks
the frames passed through).

The approach taken is promoting the current correction to be a ppi
field (like t1..t4), while killing the other "temporary" field that
was used before this commit to keep the value that was going to be added.

Note, however, that how cField is used in the back path (t3,t4) is not
clear to me, and the specification is not helping: if the correction
is automatically updated by transparent clocks in my path, the master
should send back to me the value it received for my delay request.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>