arch-wrpc: wrc_ptp_ppsi.c copied from wrpc-sw
This file comes from commit e00d99f of wrpc-sw.  It used to live there
as an alternative to wrc_ptp_noposix.c, as a top-level entry point to
the ptp engine.

However, ppsi is quite a moving target these months, and most
improvements in modularity of ppsi require a change in the top-level
file (like we change the main and startup files in the other
architectures).

Thus, this commit brings the file into ppsi (and wrpc-sw will remove
it when using this version of ppsi as submodule) so we can be work
on this repository and build-test everything with no effort, saving us
from always committing to wrpc-sw as well.

Note: this commit copies the file unchanged, and doesn't even try to
build it; later commits make the required adjustments to finally build
it.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>