arch-wrpc: major cleanup: now the spec works
This commit removes the part of wrpc.h that is not used.  It was there
when we tried to build a main function in ppsi itself, but now we
depend on the exteranl wrpc-sewpackage to provide drivers and headers,
so wrpc.h is now just for local material.

Then, wrc_ptp_ppsi.c is fixed with regard to globals and all the rest.
Moreover, I moved most of the initialization of static variables into
compile-time assignments instead of run-time, since there's nothing
dynamic here in those fields.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>