arch-gnu-linux/posix-socket.c: Fix order of EVT/GEN frames reception
This patch fixes a bug related to reception order of msgs. Sync msgs
are defined as "Event" messages and FollowUp are "General" messages, and we
need to handle them in the very same order they were sent (i.e. FollowUp is
after Sync). With the implementation prior to this patch, we tried to
serve Event and General alternatively. Now we always check for Event frames
before General ones. I didn't find, in ieee spec, whether they cleary state
that Event class is higher priority, at reception, with respect to General,
but this is likely to be in this way, since Event messages are those needing
the most precise timing.

For more information about PTP message classes, see IEEE 1588v2, pag. 17,
paragraph 6.4.

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>