servo.c: restored behaviour as it was in ptpd
This patch is temporary and is thought only to guarantee the correct
convergence of the algorithm. Some of the latest changes in servo.c, about
how time is adjusted and how the max value for observed drift is set,
made it unstable. With this patch, the behaviour is restored to the same as
ptpd, though it still needs to be reviewed to increase performance, readability
and easier maintainability.

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>