slave_only, master_only, iface_name and ethernet_mode defined as per-port cfg
Due to multilink needs, those variables are no more global but can be defined
as a different setting for each port. They will be set in a /etc/ppsi.conf file,
which is not yet handled, but will be soon.
struct pp_runtime_opts will contain only multiport settings.
If the /etc/ppsi.conf file is not found or not handled (like it is at present),
ppsi accepts and properly handles the same cmd line options as in the past.

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>