Default Data Set initialized in pp_open_instance
Default Data Set is common to all ports, so it must be defined globally.
pp_open_instance now gets a pp_globals object; the same should be done for
pp_close_intance, but is not urgent since it is currently used by wrpc arch
only, which is a single port arch.

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>