include/ppsi/ppsi.h defined GDSDEF macro
GDSDEF macro returns the defaultDS field of struct pp_globals. Actually,
it is the same pointer returned by DSDEF, since the default dataset is unique.
The DSDEF macro is not removed both to preserve commit history and to increase
readability (we do not take ppg pointer around and DSDEF(ppi) is shorter
than GDSDEF(GLOBS(ppi))).

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>