diag.c: honor DIAG_PUTS if provided
if the user passes "-DDIAG_PUTS=sth~ in USER_CFLAGS at build time, al
diagnostic messages are directed to this special puts (through sprintf
to a local buffer) instead of going to the normal printf (and thus the
default puts).

This is going to be used by wrpc-sw, to leave the default output channel
clean for the user shell.

Unfortunately, frame dumps cannot go to this special puts, as
lib/dump-funcs.c just call printf and I don't know how to fix that
without a massive change in the code.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>