Split timex adjust functions in adjust_offset and adjust_freq operations
This fixes the correct behaviour of ppsi in arch-wrpc, where adjust_freq
is not available.
Still some more fixes may be needed in servo.c. I'm not sure about the
parameter -adj to be passed both in adjust_freq and adjust_offset case.
Anyway, this has been tested on two SPECs and reaches synchronization.

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>