proto-standard/common-func.c: fix st_com_slave_handle_announce function
Since we have a call to bmc inside this function, we do not need to
unpack Announce message locally.

This patch fixes a bug with white rabbit extension, since it did not
update the foreign master record with most recent wrFlags. As a result,
at the end of wr handshaking, the slave (very often) did not handle
correctly the parentWrModeOn flag.

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>