general (and big): rename arch-gnu-linux to arch-unix
The contents of this directory are pretty portable anyways, and
the internal files were actually called "posix-io.c" and so on.
So let's get rid of the too-politica "gnu-linux" naming and of
the too-little-political "posix" at the same time.  We build
on Unix roots, so let's call our architecture like that.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>