ppsi.h: declare unix_time_ops and unix_net_ops for everyone
The unix time and network operations are going to be used as a fall-back
solution by other timing engines.

For example, hardware-specific timestamping capabilities can be
selected (such as the Linux ioctl API) but if the Ethernet interface
(and thus the pp_instance) doesn't support the feature, related code
will be able to set back the unix operations in place.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>