arch-wrs: new directory for wrs architecture
This commit is just a copy of arch-unix, as it is in commit 97c0c129.
The only difference is in file names, as they have been renamed with
wrs- prefix

arch-wrs will host the stuff for WR switch architecture. It will
be very similar to arch-unix, with the only difference of using
wr switch HAL (Hardware Abstraction Layer) instead of the standard
unix "read/write/ioctl to dev" to drive the White Rabbit specific
hardware.

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>