time-wrs: declare your functions (unix one are back static)
We need to make WR-specific things while managing the time
and network operations (even if proto-ext-whiterabbit does most
of the stuff without calling the time operations).

Thus, open-code the wrs operation, even if temporary they
just call the unix ones (which, however, are static to their
own files, and only the overall data structure is exported).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>