proto-standard: do not depend on pp_links configuration
The pp_links mechanism, currently based on /etc/ppsi.conf file, shall not
be mandatory, since not all archs will use it. Thus, this patch reduces
at its minimum the dependency on ppi->nlinks. If an arch is not using
the pp_links, it just has to define nlinks=0 in pp_globals struct, and
the engine will work as well, assuming one pp_instance has been
configured in the main.

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>