time-wrs/wrs-time.c: implement pp_time_operations
The previous implementation of wrs_time_ops was a bare call
of unix_time_ops, but this was not correct since wrs must handle
time on fpga, not the system time. wrs_calc_timeout is still using
unix_time_ops, since it just needs a monotonic timestamping, without
requiring the best precision.

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>