servo: rename some functions (no technical effect)
The servo implementation we brought forward from ptpd was strange
and difficult to understand. This commit is the first step in trying
to make some order, but it doesn't actually change anything but names.

The entry points for the servo are now called pp_servo_init(),
pp_servo_got_sync() and pp_servo_got_resp().  A two-step protocol
calls got_sync after getting the follow-up message.

Also, the hook for response messages is called handle_resp with a
name similar to the hooks for sync and follow-up.

The servo itself is not changed, but I reordered the functions so
the calculations appear in the real order in which they happen.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>