servo: minor changes (negligible technical effect)
This commit is another step in simplifying the servo. It makes the following
changes:

- format_TimeInternal is just moved (no effect)

- pp_update_offset() and __pp_update_clock() are squashed together.

They were called in sequence anyways, as two halves of a single thing.
This changes the error return paths, because the whole procedure is
aborted in case of error.  I've never seen such errors, but I think if
something is "aborted" it should not proceed.

One non-error return from the first half is turned into a goto, to avoid
changing indentation (that appears in the commit diff).

- the meaning of "ofst_first_updated" is slightly changed.

If we get an error in the sync phase the bit is cleared so delay-resp
won't use the tuple.  The field is then renamed in the next commit.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>