servo: fix correction factor and remove duplicate fields
"delay_ms" and "m_to_s_delay" were duplicates. This removes one of them
(and the same for s_to_m) and a long-standing "TODO check".

The original code was unclear about correction factor, that sometimes
was removed and sometimes not. Since the value is an overhead figure
added by any transparent clocks in the path, this commit counts it
immediately, to forget about it as soon as possible.  This also makes
the code more symmetric (t1,t2 code is similar to t3,t4 code).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>