servo: make all calculations/corrections at the end
When we only have t1/t2 we can't do all the corrections, as we miss t3/t4.

This moves all the calculations from the first servo entry point
(sync/follow-up) to the second one (delay-response).  Diagnostic
prints are moved at the end as well.

Nothing else is changed by this commit: the code is just moved.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>