servo: move more code from got_sync to got_resp (no technical effect)
This moves the first check for "max_delay" into the calculation
function. As a side effect we got rid of the field "t1_t2_valid" in
the servo structure.

While nothing is changed in the calculations (and I see the same
(mis)behaviour in my network) we now have a sane servo.c, where
pp_servo_got_sync only takes care of the correction factor for the
sync/follow-up message and then everything is done in
pp_servo_got_resp.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>