servo: change message strings (no technical change but in logs)
This reports more information about the servo, to help me trace some
misbehaviour. At level 2 it also reports T1..T4 and at level 3 it
reports the correction fields (currently always 0 in my setup).

The big hunk in this diff is just a reindentation to remove a goto
introduced a few commits ago; it is now an else stanza. Try "git show -pw"
if in doubt.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>