servo: remove needless ofm_filter
The "offset from master" filter was "a simple two-sample average" as
documented in the header, and the result of such average is used only
in the PI controller.

Thus, by averaging we have no effect at all in the I component and
only a spread over two samples of the P component for each sample.  In
practice, it has no effect at all.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>