servo: remove wrong and needless average in owd
The "Use the average between current value and previous one" in the
"one way delay" calculation was actually a running-average of two
samples. This happened before the real (and properly implemented)
running average of the value.  By removing the first one we better
respect the configured width of the filter.

As a side effect the default (64 samples) is now effective, while earlier
it behaved like a 128-sample running average. It converges faster.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>