time-wrs: fall back on unix time
When setting time, set Unix time as well. And when asking for the time,
return Unix time if the WR inter process communication fails.

Currently (and in all past history) the IPC call to get the time is
not implemented, so we are sure we fail. But I keep it in place nonetheless
in case we'll have it working in the future. Such time_get calls are
mainly for timestamping the diagnostic messages anyways.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>