white-rabbit: when T3 is invalid, don't call servo
When running a WR device as non-WR slave, we adjust the clock instead
of the phase.  When adjusting the counters, timestamps are reported as
invalid until two increments of the "seconds" register.

So, when T3 is not valid (tx stamp of delay_request), avoid running
the non-WR servo. A later iteration of the slave FSM will work.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>