arch startup: alloc max_links instances
The next commit makes a little order in startup files, and config file
parsing ends up after all allocations, so we can't reply on
ppg->nlinks to be properly set.

In hosted environments we have no issues with a few kilobytes more, so
allocate the max number of instances even if less are used.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>