wrs-socket: implemented UDP mode
This is the true and solid UDP support for WR. It manages both
WR and non-WR communication, both as master and slave.

The previous commit was wrong:

  5c91834 wrs-socket: if UDP is requested, fall back to non-WR

the reason is that we can't use Unix timestamps and WR time: to discipline
the WR mechanism we need to collect WR timestamps. So now UDP is part
of wrs-socket.c, to use the WR timestamps from hardware.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>