general: rename oneWayDelay to meanPathDelay
The 1588 standard talks extensively about "mean path delay"
and the meanPathDelay variable.   Even if the name is not the
best one, we'd better rename the fields and variable.

This is a massive change, but the currenct code has been validated at
the Lemgo plug fest, and I don't expect any need to revert older
patches at this point.  I'm sorry for git-blame users, including myself,
but this change is a step forward.  I should have knonw better when
choosing "one way delay" between the two names of the older code.

Thanks to Aurelio for noting the problem.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>