config: new config parser, to simplify further additions
This is a rewrite of the configuration parser, more based on data
structures se we'll soon be able to read more than one configuration
file (to add special cases while developing) and protocol extensions
will be able to extend the table of allowed options.

Both the configuration file format and the layout of the code calling
lib/conf.c::pp_config_file() is unchanged in this commit. This means
that even if you enable config diagnostics on the command line it
won't have effect, because the config file is parsed before the
command line. This is fixed by a later commit.

The only visible change in this commit is that "port" is allowed as a
synonim for "link", to be deprecated soon (but it will not be removed
from the parsing of the config file). Thus, pp_instance->link_name is
renamed to port_name.

The compiled size of this commit is the same as the previous one.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>