cmdline: added "-C <config>" option
This option sets configuration options from command line,
as specified in CONFIG_ITEM.
CONFIG_ITEM must be a valid configuration string, so you'll likely
need to quote it on the command line.

The semicolon can be used as a separator, to stick several
configuration items in the same argument.  This option can also be
used more than once on the same command line, to set different
parameters. If used twice to set the same parameter, the last value
overrides previous values -- and later config files, specified with
"-f" may override -C values.