bugfix: state-fault must delay a while before reinitializing ppi
This is a long-standing bug introduced when fsm.c was changed to
immediately enter a new state (before that, we had a grace period in
state-faulty).  The problem is exposed by running the suggested
config file, if eth1 and/or eth2 exist but are currently down.

This patch introduces a specific timeout, so we wait 4 seconds before
reinitializing the pp instance; we thus avoid burning 100% CPU Time in
trying again and again.

As a side effect, this reduces timeouts provided to extensions from 4
to 3. I could have reused one of the other timeouts, as faulty state
does not use them, but I'd better have proper timeout names in
diagnostics.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>