general: define CLASS names and rename a global structure
This defines PP_CLASS_SLAVE_ONLY and similar names, and uses them properly.

Also, the global "struct default_rt_opts" was not following the
conventions about naming, so rename it to be a "pp" thing.  Moreover,
since we put all globals in ppg and we have ppg->rt_opts, try to use
that instead of a separate global.  Unfortunately the initialization
sequence is not well ordered, so __pp_default_rt_opts still remains
as a global variable, double-underscored to show it's not meant to be used.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>